api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
elasticsearch[helpers]>=8.11.0,<9.0.0
sentence-transformers>=2.7.0
langchain>=0.1.16
google-generativeai>=0.3.2
python-dotenv>=1.0.1
mangum>=0.17.0
tiktoken>=0.6.0
